requirements.txt
asgiref==3.7.2
attrs==23.2.0
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
crispy-bootstrap5==2024.2
cryptography==42.0.5
defusedxml==0.7.1
Django==5.0.3
django-allauth==0.61.1
django-crispy-forms==2.1
django-phonenumber-field==7.3.0
djangorestframework==3.14.0
drf-spectacular==0.27.1
idna==3.6
inflection==0.5.1
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
oauthlib==3.2.2
phonenumberslite==8.13.32
pycparser==2.21
PyJWT==2.8.0
python3-openid==3.2.0
pytz==2024.1
PyYAML==6.0.1
referencing==0.34.0
requests==2.31.0
requests-oauthlib==1.4.0
rpds-py==0.18.0
sqlparse==0.4.4
typing_extensions==4.10.0
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.1
